move notification class to components package
