+++ sphinx/service/features/media-player/feature-service-media-player-android/src/main/java/chat/sphinx/feature_service_media_player_android/service/MediaPlayerService.kt
@@ -16,6 +16,7 @@ import chat.sphinx.concept_service_media.UserAction
 import chat.sphinx.feature_service_media_player_android.MediaPlayerServiceControllerImpl
 import chat.sphinx.feature_service_media_player_android.model.PodcastDataHolder
 import chat.sphinx.feature_service_media_player_android.service.components.AudioManagerHandler
+import chat.sphinx.feature_service_media_player_android.service.components.MediaPlayerNotification
 import chat.sphinx.feature_service_media_player_android.util.toServiceActionPlay
 import chat.sphinx.logger.SphinxLogger
 import chat.sphinx.logger.e
